Baggage Drop Zone Started
Basic systems for the baggage drop zones started with bags becoming inactive upon collision with one. Also includes a Game Controller class to handle score incrimination for the player and timing(yet to be fully implemented). A simple UI has also been added to show the players score in the game view.