Updates to UI, Pause Menu & Baggage Destruction
UI within the level updated to contain same design as found within game in a frame images. Pause Menu as well as start button added to better control when the game is in play. Objects placed at the end of conveyor belts to destroy baggage once out of sight of the player.