Camera Altered, Drop Zone Bug Fixed, End Screen Added
Camera rotation altered due to bugs with the mouse drag of object feature. Drop Zone display bug where the display bag would not match the required bag was fixed. End screen added displaying the players score at the end of the game. Additionally the NPC model on the drop zone trolleys was updated.